requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
matplotlib==3.7.2
joblib==1.3.2
python-docx
